backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
distlib==0.4.0
Django==5.2.4
django-cors-headers==4.9.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
filelock==3.19.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.7.0
packaging==25.0
platformdirs==4.4.0
postgrest==2.22.0
propcache==0.4.1
psycopg==3.2.10
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
PyMySQL==1.1.2
realtime==2.22.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
virtualenv==20.34.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.22.0
